+++ common/autoware_auto_perception_rviz_plugin/include/autoware_auto_perception_rviz_plugin/object_detection/detected_objects_display.hpp
@@ -14,7 +14,7 @@
 #ifndef OBJECT_DETECTION__DETECTED_OBJECTS_DISPLAY_HPP_
 #define OBJECT_DETECTION__DETECTED_OBJECTS_DISPLAY_HPP_
 
-#include <object_detection/object_polygon_display_base.hpp>
+#include "autoware_auto_perception_rviz_plugin/object_detection/object_polygon_display_base.hpp>
 
 #include <autoware_auto_perception_msgs/msg/detected_objects.hpp>
